requirements.txt
Django==4.2.3
djangorestframework==3.14
drf_spectacular>=0.26.4